Intial work on Bandwidth Problems with Powerline
Initial work on article about ethernet over powerline and it's
limitations and problems on an unconventional power system.

Missing parts are the conclusion and potential problems caused by
an inverter.
Additionally, I want to write about what I found on the white-paper
on AV2
https://web.archive.org/web/20170918081255/http://www.homeplug.org/media/filer_public/a1/46/a1464318-f5df-46c5-89dc-7243d8ccfcee/homeplug_av2_whitepaper_150907.pdf
I also want to extrapolate results shown in a slideshow about AV1 to
AV2.
https://web.archive.org/web/20150923143912/http://www.homeplug.org/media/filer_public/25/4f/254f6adb-096a-4913-842b-91e3775da045/devolo_presentation.pdf